requirements.txt
# =============================================================================
# Poetry Analyzer App - Requirements
# Clean requirements for Python 3.13 compatibility
# =============================================================================

# =============================================================================
# WEB FRAMEWORK
# =============================================================================
fastapi==0.129.0
uvicorn[standard]==0.34.2
python-multipart==0.0.20
jinja2==3.1.6
python-dotenv==1.1.0

# =============================================================================
# DATA VALIDATION
# =============================================================================
pydantic==2.11.0
pydantic-settings==2.8.0
pyyaml==6.0.2

# =============================================================================
# CORE NLP - ENGLISH
# =============================================================================
spacy==3.8.7
spacy-transformers==1.3.9
spacy-legacy==3.0.12

# =============================================================================
# MULTILINGUAL NLP
# =============================================================================
stanza==1.10.1

# =============================================================================
# INDIC LANGUAGES NLP
# =============================================================================
indic-nlp-library==0.92
pyiwn==0.0.5

# =============================================================================
# DEEP LEARNING & TRANSFORMERS
# =============================================================================
transformers==4.49.0
sentence-transformers==3.4.1
torch==2.10.0

# =============================================================================
# TEXT PROCESSING & METRICS
# =============================================================================
textdescriptives==2.8.2
textstat==0.7.4
syllables==1.1.5
pyphen==0.17.0

# =============================================================================
# POETRY ANALYSIS
# =============================================================================
pronouncing==0.2.0

# =============================================================================
# SENTIMENT ANALYSIS
# =============================================================================
textblob==0.19.0
vaderSentiment==3.3.2

# =============================================================================
# DATA PROCESSING
# =============================================================================
pandas==2.2.3
numpy==1.26.4

# =============================================================================
# VISUALIZATION
# =============================================================================
matplotlib==3.10.1
seaborn==0.13.2
plotly==6.0.0

# =============================================================================
# MACHINE LEARNING
# =============================================================================
scikit-learn==1.6.1
scipy==1.13.1

# =============================================================================
# TEXT COMPARISON
# =============================================================================
python-Levenshtein==0.27.3
levenshtein==0.27.3

# =============================================================================
# WEB FRAMEWORK (ALTERNATIVE)
# =============================================================================
flask==3.1.0

# =============================================================================
# UTILITIES
# =============================================================================
colorama==0.4.6
tqdm==4.67.1

# =============================================================================
# DATABASE (PostgreSQL + SQLAlchemy + Alembic Migrations)
# Use psycopg (v3) instead of psycopg2-binary for Python 3.13 compatibility
# =============================================================================
sqlalchemy>=2.0.40
alembic>=1.13.0
psycopg[binary]>=3.2.0
python-dotenv==1.1.0

# =============================================================================
# POST-INSTALLATION
# =============================================================================
# Run after installation:
#   python -m spacy download en_core_web_trf
#   python -m spacy download en_core_web_sm
#   
# For NLTK (in Python):
#   import nltk
#   nltk.download('punkt')
#   nltk.download('averaged_perceptron_tagger')
# =============================================================================
